refactor: move shared library tests to their own file (#18479)
